transfer.sh-windows/requirements.txt
clipboard==0.0.4
requests==2.11.1
progressbar2==3.18.1
